Chat/pyproject.toml
[tool.poetry]
name = "hackathonia"
version = "0.1.0"
description = ""
authors = ["Sergio León <sergio.leon@analyticalways.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
openai = "^1.20.0"
qdrant-client = {extras = ["fastembed"], version = "^1.8.2"}
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
